@@ -1,3 +1,3 @@
 module github.com/LegendaryB/go-native-messaging
 
-go 1.16
+go 1.19
